requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
EasyProcess==0.3
entrypoint2==0.2.1
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
importlib-metadata==1.6.0
jeepney==0.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
mss==5.1.0
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pkg-resources==0.0.0
protobuf==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
